[supporter-export] Make the address_mapping part of the Exporter.
This duplicates the definition of the address_mapping. One time for the
fields that the exporter provides another time for the columns that the
CSV contains. Exporters currently don't provide a list of all keys that
they offer. This might be changed at some point. Then it would be
possible to use this list for the CSV-headers.