[campaignion_bar] Do not render empty menu items
If a user has no access to the linked page, the menu item is not rendered, but the parent item is, even if it's empty. This patch adds a check to see if the parent item is an empty container linking to the frontpage, in which case it isn't rendered either.