[google_analytics] Always ensure ga() exists.
Until the tracking script is loaded ga() is defined as dummy by the
usual GA snippets. We can use the same dummy to define it even when no
snippet is loaded (eg. cookies are being blocked by a cookie bar). This
way all tracking info is queued until the script is loaded (or not).