Update for webform_confirm_email: Simplify handling of confirmed timestamps
With the upcoming version webform_confirm_email has a confirmed
timestamp instead of just a boolean value. This can be used by
campaignion_activity.

Since webform_confirm_email didn’t have timestamps for this before and
campaignion_activity did, the more accurate date for the timestamp is
stored in the activity tables.

The update hooks are bit complicated because webform_confirm_email is an
optional dependency and we only want the hook to run if
webform_confirm_email was updated recently.